# docs/swagger.yaml
basePath: /
definitions:
...
  otentikasi.DataPermintaanLogin:
    properties:
      kata_sandi:
        description: Fallback sandi untuk login retro-kompatibel
        example: admin
        type: string
      nip:
        description: Fallback NIP untuk login retro-kompatibel
        example: "2211019"
        type: string
      sandi:
        description: Kata sandi akun
        example: sandiRahasia123
        type: string
      surel:
        description: Surel akun terdaftar
        example: operator@pgn.com
        type: string
    type: object
  otentikasi.DataPermintaanLupaSandi:
    properties:
      kata_sandi:
        description: Fallback sandi baru retro-kompatibel
        example: admin
        type: string
      nip:
        description: Fallback NIP untuk reset retro-kompatibel
        example: "2211019"
        type: string
      sandiBaru:
        description: Kata sandi baru (Minimal 8 karakter)
        example: sandiBaru123
        type: string
      surel:
        description: Surel akun terdaftar
        example: operator@pgn.com
        type: string
    type: object
  otentikasi.DataPermintaanRegistrasi:
    properties:
      kata_sandi:
        description: Kata sandi fallback (Retro-kompatibilitas)
        example: admin
        type: string
      nama_lengkap:
        description: Nama lengkap pegawai (Opsional)
        example: Leader QC
        type: string
      nip:
        description: Nomor Induk Pegawai (Fallback retro-kompatibilitas)
        example: "2211019"
        type: string
      peran:
        description: 'Peran otorisasi: OPERATOR atau LEADER (Default: OPERATOR)'
        example: OPERATOR
        type: string
      sandi:
        description: Kata sandi minimal 8 karakter (Wajib jika kata_sandi kosong)
        example: sandiRahasia123
        type: string
      surel:
        description: Surel resmi pegawai (Wajib jika NIP kosong)
        example: operator@pgn.com
        type: string
    type: object
  respon.DetailKesalahan:
    properties:
      field:
        type: string
      pesan:
        type: string
    type: object
  respon.ResponStandar:
    properties:
      data: {}
      errors:
        items:
          type: string
        type: array
      kesalahan:
        items:
          $ref: '#/definitions/respon.DetailKesalahan'
        type: array
      message:
        type: string
      meta: {}
      metadata: {}
      pesan:
        type: string
      status:
        type: string
      success:
        description: Legacy fields untuk retro-kompatibilitas
        type: boolean
      sukses:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API untuk ekosistem manufaktur dan kontrol kualitas PGNServer.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PGNServer API
  version: "1.0"
paths:
...
  /api/v1/otentikasi/daftar:
    post:
      consumes:
      - application/json
      description: Mendaftarkan kredensial staf QC/Manajemen baru dengan dukungan
        retro-kompatibilitas NIP/kata_sandi
      parameters:
      - description: Data pendaftaran
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/otentikasi.DataPermintaanRegistrasi'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      summary: Pendaftaran Pengguna
      tags:
      - Otentikasi
...
  /api/v1/otentikasi/lupa-sandi:
    post:
      consumes:
      - application/json
      description: Pemulihan akun QC dengan dukungan NIP/kata_sandi fallback
      parameters:
      - description: Permohonan Reset
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/otentikasi.DataPermintaanLupaSandi'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      summary: Lupa Sandi
      tags:
      - Otentikasi
  /api/v1/otentikasi/masuk:
    post:
      consumes:
      - application/json
      description: Memberikan Token JWT untuk akses sesi internal dengan dukungan
        NIP/kata_sandi
      parameters:
      - description: Kredensial
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/otentikasi.DataPermintaanLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      summary: Akses Masuk Pengguna
      tags:
      - Otentikasi
  /api/v1/otentikasi/profil:
    get:
      description: Menyajikan data profil staf QC/Manajemen yang sedang aktif berdasarkan
        token JWT
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/respon.ResponStandar'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/respon.ResponStandar'
      security:
      - BearerAuth: []
      summary: Data Profil Sesi Pengguna
      tags:
      - Otentikasi
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"