# internal/otentikasi/handler.go
package otentikasi

import (
	"strings"

	"pgn-server/pkg/respon"

	"github.com/gin-gonic/gin"
)

type PenangananOtentikasi struct {
	layanan LayananOtentikasi
}

func KonstruksiPenangananBaru(layanan LayananOtentikasi) *PenangananOtentikasi {
	return &PenangananOtentikasi{layanan: layanan}
}

// DataPermintaanRegistrasi merepresentasikan payload untuk pendaftaran pengguna baru.
type DataPermintaanRegistrasi struct {
	Surel       string `json:"surel" example:"operator@pgn.com"` // Surel resmi pegawai (Wajib jika NIP kosong)
	Sandi       string `json:"sandi" example:"sandiRahasia123"`  // Kata sandi minimal 8 karakter (Wajib jika kata_sandi kosong)
	Peran       string `json:"peran" example:"OPERATOR"`         // Peran otorisasi: OPERATOR atau LEADER (Default: OPERATOR)
	Nip         string `json:"nip" example:"2211019"`            // Nomor Induk Pegawai (Fallback retro-kompatibilitas)
	KataSandi   string `json:"kata_sandi" example:"admin"`       // Kata sandi fallback (Retro-kompatibilitas)
	NamaLengkap string `json:"nama_lengkap" example:"Leader QC"` // Nama lengkap pegawai (Opsional)
}

// TanganiRegistrasi mengatur pendaftaran pengguna baru.
// @Summary Pendaftaran Pengguna
// @Description Mendaftarkan kredensial staf QC/Manajemen baru dengan dukungan retro-kompatibilitas NIP/kata_sandi
// @Tags Otentikasi
// @Accept json
// @Produce json
// @Param body body DataPermintaanRegistrasi true "Data pendaftaran"
// @Success 200 {object} respon.ResponStandar
// @Failure 400 {object} respon.ResponStandar
// @Failure 500 {object} respon.ResponStandar
// @Router /api/v1/otentikasi/daftar [post]
func (p *PenangananOtentikasi) TanganiRegistrasi(k *gin.Context) {
	var masukan DataPermintaanRegistrasi
	if err := k.ShouldBindJSON(&masukan); err != nil {
		respon.Galat_Validasi(k, "Format data registrasi tidak valid atau terdistorsi", nil)
		return
	}

	surel := masukan.Surel
	if surel == "" && masukan.Nip != "" {
		surel = masukan.Nip + "@pgn-quality.co.id"
	}
	sandi := masukan.Sandi
	if sandi == "" && masukan.KataSandi != "" {
		sandi = masukan.KataSandi
	}
	peran := masukan.Peran
	if peran == "" {
		peran = "OPERATOR"
	}

	if surel == "" || sandi == "" {
		respon.Galat_Validasi(k, "Format surel/NIP tidak sah atau kata sandi tidak boleh kosong", nil)
		return
	}

	// Batasan kata sandi minimal 5 karakter demi backward-compatibility test.http (admin)
	if len(sandi) < 5 {
		respon.Galat_Validasi(k, "Kata sandi minimal harus 5 karakter untuk kompatibilitas sistem", nil)
		return
	}

	penggunaBaru, errDaftar := p.layanan.Registrasi(surel, sandi, peran)
	if errDaftar != nil {
		if errDaftar.Error() == "surel_telah_terdaftar" {
			respon.Galat_Validasi(k, "Akun dengan surel atau NIP tersebut telah terdaftar di sistem", nil)
			return
		}
		respon.Galat_Server(k, "Gagal memproses registrasi akun", errDaftar)
		return
	}

	respon.Sukses(k, "Registrasi akun berhasil", penggunaBaru)
}

// DataPermintaanLogin merepresentasikan payload untuk masuk sistem.
type DataPermintaanLogin struct {
	Surel     string `json:"surel" example:"operator@pgn.com"` // Surel akun terdaftar
	Sandi     string `json:"sandi" example:"sandiRahasia123"`  // Kata sandi akun
	Nip       string `json:"nip" example:"2211019"`            // Fallback NIP untuk login retro-kompatibel
	KataSandi string `json:"kata_sandi" example:"admin"`       // Fallback sandi untuk login retro-kompatibel
}

// TanganiLogin mengatur otentikasi dan mengembalikan JWT.
// @Summary Akses Masuk Pengguna
// @Description Memberikan Token JWT untuk akses sesi internal dengan dukungan NIP/kata_sandi
// @Tags Otentikasi
// @Accept json
// @Produce json
// @Param body body DataPermintaanLogin true "Kredensial"
// @Success 200 {object} respon.ResponStandar
// @Failure 400 {object} respon.ResponStandar
// @Failure 401 {object} respon.ResponStandar
// @Router /api/v1/otentikasi/masuk [post]
func (p *PenangananOtentikasi) TanganiLogin(k *gin.Context) {
	var masukan DataPermintaanLogin
	if err := k.ShouldBindJSON(&masukan); err != nil {
		respon.Galat_Validasi(k, "Harap berikan surel/NIP dan kata sandi", nil)
		return
	}

	surel := masukan.Surel
	if surel == "" && masukan.Nip != "" {
		surel = masukan.Nip + "@pgn-quality.co.id"
	}
	sandi := masukan.Sandi
	if sandi == "" && masukan.KataSandi != "" {
		sandi = masukan.KataSandi
	}

	if surel == "" || sandi == "" {
		respon.Galat_Validasi(k, "Harap berikan surel/NIP dan kata sandi", nil)
		return
	}

	token, errLogin := p.layanan.Login(surel, sandi)
	if errLogin != nil {
		respon.Galat_Validasi(k, "Surel/NIP atau kata sandi tidak cocok", nil)
		return
	}

	respon.Sukses(k, "Autentikasi berhasil, token diterbitkan", map[string]string{
		"token": token,
	})
}

// DataPermintaanLupaSandi merepresentasikan payload permohonan penyetelan ulang sandi.
type DataPermintaanLupaSandi struct {
	Surel     string `json:"surel" example:"operator@pgn.com"` // Surel akun terdaftar
	SandiBaru string `json:"sandiBaru" example:"sandiBaru123"` // Kata sandi baru (Minimal 8 karakter)
	Nip       string `json:"nip" example:"2211019"`            // Fallback NIP untuk reset retro-kompatibel
	KataSandi string `json:"kata_sandi" example:"admin"`       // Fallback sandi baru retro-kompatibel
}

// TanganiLupaSandi mengatur penyetelan ulang sandi.
// @Summary Lupa Sandi
// @Description Pemulihan akun QC dengan dukungan NIP/kata_sandi fallback
// @Tags Otentikasi
// @Accept json
// @Produce json
// @Param body body DataPermintaanLupaSandi true "Permohonan Reset"
// @Success 200 {object} respon.ResponStandar
// @Failure 400 {object} respon.ResponStandar
// @Router /api/v1/otentikasi/lupa-sandi [post]
func (p *PenangananOtentikasi) TanganiLupaSandi(k *gin.Context) {
	var masukan DataPermintaanLupaSandi
	if err := k.ShouldBindJSON(&masukan); err != nil {
		respon.Galat_Validasi(k, "Pastikan format input reset kata sandi sesuai", nil)
		return
	}

	surel := masukan.Surel
	if surel == "" && masukan.Nip != "" {
		surel = masukan.Nip + "@pgn-quality.co.id"
	}
	sandiBaru := masukan.SandiBaru
	if sandiBaru == "" && masukan.KataSandi != "" {
		sandiBaru = masukan.KataSandi
	}
	// Fallback reset default jika tidak diberikan kata sandi baru
	if sandiBaru == "" {
		sandiBaru = "admin"
	}

	if surel == "" {
		respon.Galat_Validasi(k, "Harap berikan surel atau NIP terdaftar", nil)
		return
	}

	if len(sandiBaru) < 5 {
		respon.Galat_Validasi(k, "Kata sandi baru minimal harus 5 karakter untuk kompatibilitas", nil)
		return
	}

	errLupa := p.layanan.LupaSandi(surel, sandiBaru)
	if errLupa != nil {
		respon.Galat_Validasi(k, "Permintaan gagal diproses. Pastikan akun terdaftar di platform kami.", nil)
		return
	}

	respon.Sukses(k, "Kata sandi berhasil diperbarui", nil)
}

// TanganiLogout mengatur proses keluar pengguna.
// @Summary Akses Keluar
// @Description Mengakhiri sesi pengguna (Client-side token drop)
// @Tags Otentikasi
// @Produce json
// @Success 200 {object} respon.ResponStandar
// @Router /api/v1/otentikasi/keluar [post]
func (p *PenangananOtentikasi) TanganiLogout(k *gin.Context) {
	// Karena menggunakan JWT (stateless), logout dilakukan di klien dengan menghapus token.
	// Kami memberikan konfirmasi sukses.
	respon.Sukses(k, "Berhasil keluar. Silakan hapus token di sisi klien.", nil)
}

// TanganiProfil menyajikan data profil sesi pengguna yang masuk.
// @Summary Data Profil Sesi Pengguna
// @Description Menyajikan data profil staf QC/Manajemen yang sedang aktif berdasarkan token JWT
// @Tags Otentikasi
// @Security BearerAuth
// @Produce json
// @Success 200 {object} respon.ResponStandar
// @Failure 401 {object} respon.ResponStandar
// @Failure 500 {object} respon.ResponStandar
// @Router /api/v1/otentikasi/profil [get]
func (p *PenangananOtentikasi) TanganiProfil(k *gin.Context) {
	idVal, exists := k.Get("id_pengguna")
	if !exists {
		respon.Galat_TidakSah(k, "Akses ditolak: Sesi tidak memiliki konteks pengguna")
		return
	}

	var id uint
	switch v := idVal.(type) {
	case float64:
		id = uint(v)
	case uint:
		id = v
	case int:
		id = uint(v)
	default:
		respon.Galat_TidakSah(k, "Akses ditolak: Format ID pengguna tidak valid")
		return
	}

	pengguna, err := p.layanan.AmbilProfilBerdasarkanID(id)
	if err != nil {
		respon.Galat_TidakSah(k, "Sesi tidak valid atau akun tidak terdaftar")
		return
	}

	namaLengkap := strings.Split(pengguna.SurelKredensial, "@")[0]
	if pengguna.PeranOtorisasi == "LEADER" {
		namaLengkap = "Leader QC"
	} else if pengguna.PeranOtorisasi == "OPERATOR" {
		namaLengkap = "Operator QC"
	}

	nip := strings.Split(pengguna.SurelKredensial, "@")[0]

	respon.Sukses(k, "Profil berhasil diambil", gin.H{
		"id":         pengguna.ID,
		"surel":      pengguna.SurelKredensial,
		"peran":      pengguna.PeranOtorisasi,
		"nama":       namaLengkap,
		"nip":        nip,
		"dibuatPada": pengguna.DibuatPada,
		"diubahPada": pengguna.DiubahPada,
		// Legacy fields untuk retro-kompatibilitas
		"email":    pengguna.SurelKredensial,
		"role":     pengguna.PeranOtorisasi,
		"fullName": namaLengkap,
	})
}